pyproject.toml
[project]
name = "beauty-city-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "environs>=14.5.0",
    "Pillow>=10.0.0",
    "django-crispy-forms>=2.0",
    "django-cleanup>=9.0.0",
    "django-phonenumber-field>=7.0.0",
    "phonenumbers>=8.13.0",
]
